Move conformance to test_utils
